% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hanley_mcneil_var}
\alias{hanley_mcneil_var}
\title{Hanley-McNeil Variance of AUC}
\usage{
hanley_mcneil_var(auc, n_cases, n_controls)
}
\arguments{
\item{auc}{Estimated AUC.}

\item{n_cases}{Number of diseased individuals.}

\item{n_controls}{Number of non-diseased individuals.}
}
\value{
Numeric variance estimate.
}
\description{
Computes the variance of the area under the ROC curve using the
Hanley-McNeil approximation.
}
\references{
Hanley JA, McNeil BJ (1982). The meaning and use of the area under a
receiver operating characteristic (ROC) curve. \emph{Radiology}
143:29-36.
}
\keyword{internal}
