% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.dtasamplesize}
\alias{print.dtasamplesize}
\title{Print Method for dtasamplesize Objects}
\usage{
\method{print}{dtasamplesize}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"dtasamplesize"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for dtasamplesize Objects
}
